[client] move OffsetsInitializer to fluss-client
